Remove impl From<VVal> for NVec in favor of VVal::nvec
